Change 0
Here I created all of my floats which are the (x,y) coordinates of my 5
rectangles. (lines 6-17) I made all of the rectangles the same size,
just their direction is slightly carried. Their x and y values are
randomly increased based on what I put as the highest number for them
to be raised by. (For example: lines 37,38). Lastly, I set it so that
if my rectangle goes off of the screen, it is randomly replaced
anywhere on the screen. (For example: lines 40-48). I also added a
background so that the only the current version and placement of the
rectangles are visible.