Change 2
I decided to make more rectangles, and to have rectangles which all go
in different directions so it seems more random. Because I added so
many rectangles, I had to assign and use a lot more variables, and to
organize my code, I added 2 voids. A Positive void which has all of the
rectangles with positive X values. Then a Negative void which has all
of the rectangles with a negative X value. Due to making some negative
rectangles, I had to change some of the other values about if it goes
off of the screen for it to reappear in a random spot. Another thing I
changed was how much the rectangles position could increase by every
time the program was run. I cut all of my numbers in half to make the
rectangles move slower.